Incomplete: pulling the update/remove action into the Dialog.
The tests have not yet matched the refactoring, but the idea is
to have the extension manager dialog trigger the update/remove
functionality, rather than the individual views.
